src/design3/decoder/ppp_common/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ppp_common"
version = "0.1"
requires-python = ">=3.11"

dependencies = [
    # if changed also update in `../Dockerfile`
    "loguru==0.7.3",
    "pymonocypher==4.0.2.5",
    "pyelftools==0.31"
]

[project.optional-dependencies]
dev = [
    "ruff"
]
